pyproject.toml
[tool.poetry]
name = "bert-for-ukranian-ner"
version = "0.1.0"
description = "This repository contains I part of my bachelor's project"
authors = ["kryvokhyzha <kriwohizha@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tokenize_uk = "0.2.0"
catalyst = "21.04.02"
transformers = "4.1.1"
tensorboard = "^2.5.0"
PyYAML = "5.3.1"
pandas = "^1.2.4"
joblib = "^1.0.1"
scikit-learn = "0.23.2"

[tool.black]
line-length = 90
target-version = ["py38"]
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 90
lines_after_imports = 2

[tool.nbqa.mutate]
black = 1
isort = 1

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.poetry.dev-dependencies]
pre-commit = "^2.9.3"
flake8 = "3.8.4"
nbqa = "^0.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
